Develop

+++ src/test/java/com/project/cinemory/domain/review/service/ReviewRatingFallbackTest.java
@@ -15,6 +15,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -54,7 +55,7 @@ class ReviewRatingFallbackTest {
         return movieRepository.save(Movie.builder().tmdbId(tmdbId).title(title).build());
     }
 
-    private WatchRecordCreateRequest watchRecordRequest(Long movieId, Double rating) {
+    private WatchRecordCreateRequest watchRecordRequest(Long movieId, BigDecimal rating) {
         return new WatchRecordCreateRequest(movieId, null, null, null, null, rating, null);
     }
 
@@ -63,7 +64,7 @@ class ReviewRatingFallbackTest {
         User user = createUser("fallback-rep@test.com", "대표기록유저");
         Movie movie = createMovie(90_001L, "대표기록 테스트 영화");
 
-        watchRecordService.addWatchRecord(user.getId(), watchRecordRequest(movie.getId(), 8.0));
+        watchRecordService.addWatchRecord(user.getId(), watchRecordRequest(movie.getId(), BigDecimal.valueOf(8.0)));
         reviewService.writeReview(user.getId(), movie.getId(), new ReviewWriteRequest("좋은 영화였다"));
 
         Optional<ReviewResponse> myReview = reviewService.getMyReview(user.getId(), movie.getId());
@@ -78,7 +79,7 @@ class ReviewRatingFallbackTest {
         Movie movie = createMovie(90_002L, "재관람 테스트 영화");
 
         // 1회차 — 별점 8.0으로 관람, 이 시점엔 대표 기록
-        watchRecordService.addWatchRecord(user.getId(), watchRecordRequest(movie.getId(), 7.0));
+        watchRecordService.addWatchRecord(user.getId(), watchRecordRequest(movie.getId(), BigDecimal.valueOf(7.0)));
         // 2회차 — 별점 없이 재관람, addWatchRecord는 항상 신규 기록을 대표로 승격시킨다
         watchRecordService.addWatchRecord(user.getId(), watchRecordRequest(movie.getId(), null));
